refactor(contracts): use single backers manager hook
