+++ src/app/collective-rewards/allocations/hooks/index.ts
@@ -1,5 +1,4 @@
-export * from './useBackerTotalAllocation'
-export * from './useGetVotingPower'
 export * from './useActivatedBuildersWithGauge'
 export * from './useBuildersWithBackerRewardPercentage'
 export * from './useCanManageAllocations'
+export * from './useGetVotingPower'
