+++ src/app/collective-rewards/rewards/builders/EstimatedRewards.tsx
@@ -8,14 +8,13 @@ import {
   Token,
   TokenMetricsCardRow,
   useGetPerTokenRewards,
-  useGetTotalPotentialReward,
 } from '@/app/collective-rewards/rewards'
 import { useBuilderContext } from '@/app/collective-rewards/user'
 import { isBuilderRewardable, useHandleErrors } from '@/app/collective-rewards/utils'
 import { withSpinner } from '@/components/LoadingSpinner/withLoadingSpinner'
 import { WeiPerEther } from '@/lib/constants'
 import { usePricesContext } from '@/shared/context/PricesContext'
-import { useReadBuilderRegistry, useReadGauges } from '@/shared/hooks/contracts'
+import { useReadBackersManager, useReadBuilderRegistry, useReadGauges } from '@/shared/hooks/contracts'
 import { FC, useEffect, useMemo, useState } from 'react'
 import { Address } from 'viem'
 
@@ -48,7 +47,9 @@ const TokenRewards: FC<TokenRewardsProps> = ({ builder, gauge, token: { id, symb
     data: totalPotentialRewards,
     isLoading: totalPotentialRewardsLoading,
     error: totalPotentialRewardsError,
-  } = useGetTotalPotentialReward()
+  } = useReadBackersManager({
+    functionName: 'totalPotentialReward',
+  })
   const {
     data: rewardShares,
     isLoading: rewardSharesLoading,
