+++ src/app/collective-rewards/rewards/hooks/index.ts
@@ -6,5 +6,4 @@ export * from './useGetLastCycleDistribution'
 export * from './useGetNotifyRewardLogs'
 export * from './useGetPerTokenRewards'
 export * from './useGetRewardDistributionFinishedLogs'
-export * from './useGetTotalPotentialRewards'
 export * from './useIsBuilderOrBacker'
