+++ src/app/collective-rewards/rewards/hooks/useGetPerTokenRewards.ts
@@ -1,10 +1,10 @@
-import { BackersManagerAbi } from '@/lib/abis/v2/BackersManagerAbi'
+import { type BackersManagerAbi } from '@/lib/abis/v2'
 import { useReadRewardDistributor } from '@/shared/hooks/contracts'
 import { UseReadContractReturnType } from 'wagmi'
 
 export const useGetPerTokenRewards = (): Record<
   'rif' | 'rbtc',
-  UseReadContractReturnType<typeof BackersManagerAbi, 'rewardsERC20' | 'rewardsCoinbase'>
+  UseReadContractReturnType<BackersManagerAbi, 'rewardsERC20' | 'rewardsCoinbase'>
 > => ({
   rif: useReadRewardDistributor({ functionName: 'defaultRewardTokenAmount' }),
   rbtc: useReadRewardDistributor({ functionName: 'defaultRewardCoinbaseAmount' }),
