+++ src/app/collective-rewards/shared/hooks/useGetABI.ts
@@ -1,4 +1,3 @@
-import { useBackerTotalAllocation } from '@/app/collective-rewards/allocations/hooks'
 import { useGetBackersRewardPercentage } from '@/app/collective-rewards/rewards'
 import { useGetCyclePayout, useGetEstimatedBackersRewardsPct } from '@/app/collective-rewards/shared'
 import { RequiredBuilder } from '@/app/collective-rewards/types'
@@ -6,7 +5,7 @@ import { useGetBuildersByState } from '@/app/collective-rewards/user'
 import Big from '@/lib/big'
 import { WeiPerEther } from '@/lib/constants'
 import { usePricesContext } from '@/shared/context/PricesContext'
-import { useReadGauges } from '@/shared/hooks/contracts'
+import { useReadBackersManager, useReadGauges } from '@/shared/hooks/contracts'
 import { useMemo } from 'react'
 import { Address } from 'viem'
 
@@ -46,7 +45,10 @@ export const useGetRewardsAbi = (backer: Address) => {
     data: backerTotalAllocation,
     isLoading: backerTotalAllocationLoading,
     error: backerTotalAllocationError,
-  } = useBackerTotalAllocation(backer)
+  } = useReadBackersManager({
+    functionName: 'backerTotalAllocation',
+    args: [backer],
+  })
   const { cyclePayout, isLoading: cyclePayoutLoading, error: cyclePayoutError } = useGetCyclePayout()
 
   const rewardsPerStRif = useMemo(() => {
